Refactor TruncateFd

+++ bindings-chasm/src/commonMain/kotlin/module/emscripten/function/SyscallFtruncate64.kt
@@ -21,7 +21,7 @@ internal class SyscallFtruncate64(
     override val function: HostFunction = { args ->
         val result: Int = handle.execute(
             args[0].asInt(),
-            args[1].asLong().toULong(),
+            args[1].asLong(),
         )
         listOf(Value.Number.I32(result))
     }
